Refactor Resource modal component (ChooseResource)
This refactor integrates the desktop and mobile component code for
the Resource modal under a common directory and consistent approach.
Previously, the mobile component was under the directory
src/components/ResourceMenu.
