--- src/hooks/selectors.js
@@ -1,1 +1,0 @@
-export const isResourceMenuShownSelector = state => state.isResourceMenuShown;
